✨ feat: Adiciona filtro savgol
