+++ code/util/class_control_panel.py
@@ -1,3 +1,7 @@
+import pandas as pd
+from scipy.signal import savgol_filter
+from typing import List, Tuple
+
 class ControlPanel():
     """É responsável por definir parâmetros para fazer diferentes
     experimentos
@@ -8,7 +12,8 @@ class ControlPanel():
                  is_grid_search: bool = False,
                  use_validation_data: bool = False,
                  use_optuna: bool = False,
-                 number_units_validation: int = 0) -> None:
+                 number_units_validation: int = 0,
+                 use_savgol_filter: bool = False) -> None:
         """É responsável por definir parâmetros para fazer diferentes
         experimentos.
 
@@ -37,3 +42,20 @@ class ControlPanel():
         self.use_validation_data = use_validation_data
         self.use_optuna = use_optuna
         self.number_units_validation = number_units_validation
+        self.use_savgol_filter = use_savgol_filter
+
+    def apply_use_savgol_filter(self,
+                                df_data: pd.DataFrame,
+                                ignore_column: List[str],
+                                window_length: int = 15,
+                                polyorder: int = 2) -> Tuple[pd.DataFrame, List[str]]:
+        new_columns = []
+        if self.use_savgol_filter:
+            
+            for column in df_data.columns:
+                if column not in ignore_column:
+                    df_data[f'{column}_savgol'] = savgol_filter(df_data[column],
+                                                                window_length,
+                                                                polyorder)
+                    new_columns.append(f'{column}_savgol')
+        return df_data, new_columns
